Fixed #58 and moved the BrowseSpeakers to Speakers/Index
