+++ src/MoreSpeakers.Web/Areas/Identity/Pages/Account/_RegisterStep5.cshtml
@@ -228,7 +228,7 @@
                     <i class="bi bi-pencil me-2"></i>
                     Complete My Profile
                 </a>
-                <a href="@Url.Page("/BrowseSpeakers")" class="btn btn-outline-primary btn-lg px-4 py-3">
+                <a asp-page="~/Speakers/Index" class="btn btn-outline-primary btn-lg px-4 py-3">
                     <i class="bi bi-people me-2"></i>
                     Explore Community
                 </a>
