+++ src/MoreSpeakers.Web/Pages/About.cshtml
@@ -172,7 +172,7 @@
                         }
                         else
                         {
-                            <a asp-page="/BrowseSpeakers" class="btn btn-light btn-lg">
+                            <a asp-page="/Speakers/Index" class="btn btn-light btn-lg">
                                 <i class="bi bi-search me-2"></i>Browse Speakers
                             </a>
                             <a asp-page="/Profile/Index" class="btn btn-outline-light btn-lg">
