+++ src/MoreSpeakers.Web/Pages/Profile/View.cshtml
@@ -9,7 +9,7 @@
     <nav aria-label="breadcrumb" class="mb-4">
         <ol class="breadcrumb">
             <li class="breadcrumb-item"><a href="/" class="text-decoration-none">Home</a></li>
-            <li class="breadcrumb-item"><a asp-page="/BrowseSpeakers" class="text-decoration-none">Browse Speakers</a></li>
+            <li class="breadcrumb-item"><a asp-page="~/Speakers/Index" class="text-decoration-none">Browse Speakers</a></li>
             <li class="breadcrumb-item active" aria-current="page">@Model.ProfileUser?.FullName</li>
         </ol>
     </nav>
