+++ src/MoreSpeakers.Web/Pages/Speakers/Index.cshtml
@@ -1,5 +1,5 @@
 @page
-@model MoreSpeakers.Web.Pages.BrowseSpeakersModel
+@model MoreSpeakers.Web.Pages.Speakers.IndexModel
 @{
     ViewData["Title"] = "Browse Speakers";
 }
@@ -26,7 +26,7 @@
 
 <!-- Search and Filter Section -->
 <div class="border my-3 p-3 bg-body-tertiary">
-    <form method="get" hx-get="/BrowseSpeakers" hx-target="#speakers-container" hx-indicator="#loading" hx-swap="morph" hx-trigger="keyup changed delay:500ms from:#searchTerm, change from:#speakerTypeFilter, change from:#expertiseFilter, change from:#sortBy">
+    <form method="get" hx-get="/Speakers/Index" hx-target="#speakers-container" hx-indicator="#loading" hx-swap="morph" hx-trigger="keyup changed delay:500ms from:#searchTerm, change from:#speakerTypeFilter, change from:#expertiseFilter, change from:#sortBy">
         <div class="row">
             <div class="col-md-6 mb-3">
                 <label for="searchTerm" class="form-label">Search</label>
@@ -71,7 +71,7 @@
                         <strong>@Model.TotalCount</strong> speaker(s) found
                         @if (!string.IsNullOrEmpty(Model.SearchTerm) || Model.SpeakerTypeFilter.HasValue || Model.ExpertiseFilter.HasValue)
                         {
-                            <a asp-page="/BrowseSpeakers" class="btn btn-outline-secondary btn-sm ms-2">
+                            <a asp-page="/Speakers/Index" class="btn btn-outline-secondary btn-sm ms-2">
                                 <i class="bi bi-x-circle"></i> Clear Filters
                             </a>
                         }
@@ -138,7 +138,7 @@
         if (!string.IsNullOrEmpty(Model.SortBy))
             queryParams.Add($"SortBy={Model.SortBy}");
 
-        return $"/BrowseSpeakers?{string.Join("&", queryParams)}";
+        return $"/Speakers/Index?{string.Join("&", queryParams)}";
     }
 
 }
